added 7th sem & saving of grades as text for re-use.
Added a text box up-top, where entering the grades in order separated by blank spaces, and clicking the calculate button automatically assigns the values to the drop downs and calculates each semester's GPA. Works on for 7th Sem as of now (both lateral and regular).

To get this text you can enter all grades up to seventh semester and then calculate the CGPA, this will by default copy the grade-text to your clipboard. It is also displayed and can be copied from there too.