Gamma Version - 8th Semester grades, EDP grade added, Semester wise saving capability, Improved grade input
Safe to say this is the final form of the CGPA calculator. Added the entries for 8th Sem Grade, and also for the missing EDP grade in 6th semester.

Grades can be copied as text semester-wise, so all batches can benefit from the calculator.

The input grades are assigned to the entries (drop-downs) from cascaded if-statement blocks, so we calculate only necessary semesters.

Everything is optimized for lateral students also.

Spoiler: Added a goodbye message to the 8th semester students @ line 1697/1711