IMPORTANT: Corrected CGPA calculation for 8th Semester
There was a mistake in the code for the calculation of CGPA for Regular Students for the 8th Semester. It was brought to my notice by Jagannathan and Sudhagar as the CGPA differed from the one on their provisional marksheets. It has been corrected. My sincere apologies for the same, even I was elated to find even my grade boosted.